pyproject.toml
[project]
name = "perfx"
version = "0.1.0"
description = "A configurable performance evaluation framework"
authors = [
    {name = "Jianhong Zhao", email = "zhaojianhong96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "matplotlib>=3.5",
    "numpy>=1.21",
    "pandas>=1.3",
    "click>=8.0",
    "rich>=12.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=0.950",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]

[tool.coverage.run]
source = ["perfx"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[project.scripts]
perfx = "perfx.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=24.8.0",
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.1",
]

[tool.hatch.build.targets.wheel]
packages = ["perfx"]

[tool.hatch.build.targets.wheel.sources]
"perfx" = "perfx" 
